pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "mcp[cli]>=1.8.0",
    "openai>=1.0.0",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
]

[tool.uv]
index-url = "https://pypi.org/simple/"
